refactoring, changed minkore to kore
