+++ pom.xml
@@ -5,7 +5,7 @@
   <modelVersion>4.0.0</modelVersion>
 
   <groupId>org.kframework.k</groupId>
-  <artifactId>minikore</artifactId>
+  <artifactId>kore</artifactId>
   <version>1.0-SNAPSHOT</version>
   <packaging>jar</packaging>
