+++ src/main/scala/org/kframework/kore/parser/TextToMini.scala
@@ -1,15 +1,16 @@
-package org.kframework.minikore.parser
+package org.kframework.kore.parser
 
 import org.apache.commons.lang3.StringEscapeUtils
-
-import org.kframework.minikore.interfaces.outer._
-import org.kframework.minikore.interfaces.pattern._
+import org.kframework.kore.interfaces.builders.OuterBuilders
+import org.kframework.kore.interfaces.builders.PatternBuilders
+import org.kframework.kore.interfaces.outer._
+import org.kframework.kore.interfaces.pattern._
 
 
 /** Parsing error exception. */
 case class ParseError(msg: String) extends Exception(msg) // ParseError.msg eq Exception.detailMessage, i.e., msg() == getMessage()
 
-/** A parser for [[org.kframework.minikore.interfaces.pattern]].
+/** A parser for [[org.kframework.kore.interfaces.pattern]].
   *
   * @constructor Creates a new parser.
   */
@@ -19,13 +20,13 @@ class TextToMini(ob: OuterBuilders, pb: PatternBuilders) {
 
   private val scanner = new Scanner()
 
-  /** Parses the file and returns [[org.kframework.minikore.implementation.outer.Definition]]. */
+  /** Parses the file and returns [[org.kframework.kore.implementation.DefaultOuter.Definition]]. */
   @throws(classOf[ParseError])
   def parse(file: java.io.File): Definition = {
     parse(io.Source.fromFile(file))
   }
 
-  /** Parses from the stream and returns [[org.kframework.minikore.implementation.outer.Definition]]. */
+  /** Parses from the stream and returns [[org.kframework.kore.implementation.DefaultOuter.Definition]]. */
   @throws(classOf[ParseError])
   def parse(src: io.Source): Definition = {
     try {
